requirements.txt
torch>=2.0
torchvision>=0.15
Pillow>=9.0
numpy>=1.24
matplotlib>=3.7
tqdm>=4.65
imageio>=2.31
